Move template classes to utils
